+++ packages/widgets/src/widgets/lagaufre/main.ts
@@ -9,6 +9,12 @@ const widgetName = "lagaufre";
 
 type Service = {
   name: string;
+  /**
+   * String to use as aria-label if the visible name is not self-explanatory enough.
+   * It should start with the visible name or at least include it
+   * (see https://www.w3.org/WAI/WCAG22/Understanding/label-in-name.html)
+   */
+  accessibleName?: string;
   url: string;
   maturity?: string;
   logo?: string;
@@ -233,7 +239,9 @@ listenEvent(widgetName, "init", null, false, async (args: GaufreWidgetArgs) => {
 
       anchor.setAttribute(
         "aria-label",
-        service.name + (service.maturity ? ` (${service.maturity})` : "") + (args.newWindowLabelSuffix || ""),
+        (service.accessibleName || service.name) +
+          (service.maturity ? ` (${service.maturity})` : "") +
+          (args.newWindowLabelSuffix || ""),
       );
       anchor.href = service.url;
       img.src = service.logo;